REF: Add spatial utils module for pylint
